updated v1.08

+++ read_file.cpp
@@ -1,6 +1,9 @@
 #include <iostream>
 using namespace std;
 
+//#include <opencv2/highgui.hpp>
+//using namespace cv;
+
 int main()
 {
     FILE *pf = fopen("beans_text1.txt", "rb");
@@ -9,8 +12,8 @@ int main()
     lsize = ftell(pf);
     rewind(pf);
 
-    uint16_t *buffer;
-    buffer = (uint16_t *)malloc(sizeof(uint16_t) * lsize);
+    uint8_t *buffer;
+    buffer = (uint8_t *)malloc(sizeof(uint8_t) * lsize);
 
     size_t result;
     result = fread(buffer, 1, lsize, pf);
@@ -19,7 +22,9 @@ int main()
     int num_rows = 309;
     int num_cols = 116;
 
-//    Mat img;
+    
+
+//    Mat img[96];
 //    imshow("image", buffer);
 //    waitKey(0);
 }
